added HomeScreen route
